move the source files up
