+++ service/search_proxy_test.go
@@ -1,7 +1,7 @@
 package service
 
 import (
-	"github.com/splunk/ssc-client-go/lib/model"
+	"github.com/splunk/ssc-client-go/model"
 	"testing"
 	"time"
 )
