+++ service/search_test.go
@@ -2,7 +2,7 @@ package service
 
 import (
 	"testing"
-	"github.com/splunk/ssc-client-go/lib/model"
+	"github.com/splunk/ssc-client-go/model"
 	"github.com/stretchr/testify/assert"
 	"time"
 )
